+++ src/L293_std.hpp
@@ -17,8 +17,8 @@ class L293 : public L293_base
 			void forceStop( uint16_t handlingTime );	///< Stops the motor by electrically braking it
 			bool isForceStopped() const ;						///< Tells the information about the motor that the name says
 
-			void forward( uint8_t _PWMDC = 0 ) override ; ///< Makes the motor to go forward and sets a new speed value
-			void back( uint8_t _PWMDC = 0 )		 override ; ///< Makes the motor to go reverse and sets a new speed value
+			void forward( uint16_t _PWMDC = 0 ) override ; ///< Makes the motor to go forward and sets a new speed value
+			void back( uint16_t _PWMDC = 0 )		 override ; ///< Makes the motor to go reverse and sets a new speed value
 			bool isForward() const override ;
 			bool isReverse() const override ;
 			bool isStopped() const override ;
